CR: Rework default page number
