+++ src/app/_constants/paginationConstants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_PAGE_NUMBER = 1
