+++ src/app/_hooks/usePagination.ts
@@ -1,11 +1,30 @@
 import { NextServerSearchParams } from '@/types/searchParams'
 
-import { validatePageNumber } from '@/utils/validatePageNumber'
+import { DEFAULT_PAGE_NUMBER } from '@/constants/paginationConstants'
 
 type UsePaginationProps = {
   totalEntries: number
   entriesPerPage: number
   pageQuery?: NextServerSearchParams[string]
+}
+
+export function validatePageNumber(
+  pageQuery: string | undefined,
+  pageCount: number,
+): number {
+  const pageQueryNumber = Number(pageQuery)
+  const isValidNumber = Number.isInteger(pageQueryNumber) && pageQueryNumber > 0
+  const isWithinRange = pageQueryNumber <= pageCount
+
+  if (!isValidNumber) {
+    return DEFAULT_PAGE_NUMBER
+  }
+
+  if (!isWithinRange) {
+    return pageCount
+  }
+
+  return pageQueryNumber
 }
 
 export function usePagination({
